requirements.txt
Django==4.2
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
gunicorn==22.0.0
whitenoise==6.8.2
requests==2.32.3
django-cors-headers==3.14.0
pydantic==2.10.4
django-filter==24.3
django-storages==1.14.4
redis==5.2.1
django-redis==5.4.0
channels==4.2.0
channels-redis==4.2.1
uvicorn[standard]==0.34.0
amqp==5.3.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-extensions==4.1